Made changes to snake animation
Created an if statement in the paint animation in the snakeObject class. This if statement checks to see if the previous snakeSegement in the array is in the same direction as the current snakeSegment in the array. If it is not, it will be put in the same direction as the previous snakeSegment.